renames box_plot to plot
