+++ axelrod/plot.py
