+++ axelrod/tests/test_plot.py
